refactor(web): split day calendar grid responsibilities
